requirements.txt
# AI Productivity Platform Requirements
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0

# Async Task Processing
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Caching
django-redis==5.4.0

# AI Integration
openai==1.3.8

# Authentication & Security
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0

# Image Processing
Pillow==10.1.0

# HTTP Requests
requests==2.31.0

# Data Processing
python-dateutil==2.8.2

# Development Tools (optional, remove in production)
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Monitoring & Logging
sentry-sdk[django]==1.38.0

# API Documentation
drf-spectacular==0.26.5

# File Uploads
django-storages==1.14.2
boto3==1.34.0  # For AWS S3 support

# Email
django-anymail==10.2

# Background Jobs Monitoring
flower==2.0.1

# Performance
django-cachalot==2.6.1

# Security
django-ratelimit==4.1.0

# Utilities
pytz==2023.3
six==1.16.0

# Health Checks
django-health-check==3.17.0

# Data Validation
jsonschema==4.20.0

# Testing (development only)
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality (development only)
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Documentation (development only)
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0